Remove direct link to invent.kde.org.
Linking to raw files on invent.kde.org is *strictly* and *absolutely* prohibited in the strongest possible terms and is something that Sysadmin does not permit in any form.
This is particularly the case when those files will be fetched by end user systems, as it creates the risk of an unintentional Distributed Denial of Service attack on invent.kde.org.

CCMAIL: sysadmin@kde.org
CCMAIL: kde-core-devel@kde.org
CCMAIL: community-wg@kde.org